requirements.txt
pyyaml
jinja2
luqum